requirements.txt
pandas==0.25.1
numpy==1.17.2
scipy==1.2.1
Django==2.2.7
djangorestframework==3.11.0
psycopg2==2.8.4
pystan==2.17.1.0
holidays==0.9.12
matplotlib==3.1.2
scikit-learn==0.21.2
xgboost==0.90
kafka-python==1.4.7
django-cors-headers==3.2.0
boto3==1.10.45
django-storages==1.8
python-dateutil==2.8.1
django_compressor
avro-validator==1.0.6
requests==2.23.0
django-mptt==0.11.0
pydruid==0.5.8
django-filter==2.3.0
fbprophet==0.6

